fix: Resolve the flickering issue with swipeCards - The flickering issue seems to have been resolved without significant changes. - Root cause is unclear, and the issue may reappear in the future.
