+++ client/src/components/Swipe/SwipeDeck.jsx
@@ -1,9 +1,9 @@
-import {useEffect, useRef} from "react";
+import {useEffect, useRef, useState} from "react";
 import {animated} from "@react-spring/web";
 
-import useSwipeDeckManager from "./hooks/useSwipeDeckManager";
 import SwipeDebugInfo from "./SwipeDebugInfo";
 import SwipeCardContent from "./swipeCard/SwipeCardContent";
+import useSwipeAnimationHandler from "./hooks/useSwipeAnimationHandler.js";
 
 const FIRST_CARD_INDEX = 0;
 const MAX_Z_INDEX = 100;
@@ -13,15 +13,29 @@ const SwipeDeck = ({users, setUsers, deckWidth, onSwipe}) => {
     const swipeCardRef = useRef(null);
     const cardImageRef = useRef(null);
 
-    const {bind, swipeDebugInfo, animatedStyles, y} =
-        useSwipeDeckManager(setUsers, deckWidth, cardImageRef, onSwipe);
+    const onSwipeOut = (direction) => {
+        setUsers((prevCards) => {
+            console.log("BEFORE THE SWIPE!");
+            console.log(prevCards);
+            onSwipe(prevCards[0].id,direction);
+            return prevCards.slice(1)
+        });
+    };
+
+    const { bind, swipeDebugInfo, reset: resetAnimation, animatedStyles, y } = useSwipeAnimationHandler(deckWidth, onSwipeOut, cardImageRef);
 
     useEffect(() => {
         console.log("Component rendered or re-rendered");
     });
 
+    useEffect(() => {
+        resetAnimation();
+    }, [users]);
+
+
+
     const renderCards = () =>
-        users.map((user, index) => (
+        users.map((user, index, arr) => (
             <animated.div
                 ref={index === FIRST_CARD_INDEX ? swipeCardRef : null}
                 key={index}
@@ -30,7 +44,7 @@ const SwipeDeck = ({users, setUsers, deckWidth, onSwipe}) => {
                     touchAction: "none",
                     position: "absolute",
                     zIndex:
-                        index === FIRST_CARD_INDEX ? MAX_Z_INDEX : users.length - index,
+                        index === FIRST_CARD_INDEX ? MAX_Z_INDEX : arr.length - index,
                     ...(index === FIRST_CARD_INDEX ? animatedStyles : {}),
                 }}
             >
